steamcompmgr: Fix condition on enabling mouse cursor constraint
When the underlying application sets a cursor constraint e.g. for the
purpose of relative mouse mode, Gamescope tells the host compositor to
lock the cursor. When to do this is currently determined in
MouseCursor::UpdatePosition by checking if there is an active
constraint with a mouse cursor position hint.

However, an active constraint without a mouse cursor position hint
should be sufficient to lock the cursor. Therefore, fix the logic
to lock the cursor without requiring a position hint.

This fixes relative mouse motion while rotating the camera in World of
Warcraft. Without this fix, the cursor moves while in relative motion
mode when it shouldn't.